% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backups.R
\name{backup_file_path}
\alias{backup_file_path}
\title{Find file in backup}
\usage{
backup_file_path(backup, file, domain = "")
}
\arguments{
\item{backup}{An \code{ios_backup} object (or something that can be passed
to \code{get_backup}).}

\item{file}{The relative name of the file you want to find.}

\item{domain}{The domain in which to search for the file.}
}
\value{
For each input, will return a character value with the
 true path to the file on disk or NA if the location cannot be
 determined (no entry in the manifest)
}
\description{
Return the path for a specific file in the backup
}
\details{
This function translates the human-readable relative file
paths and domains into the fileID necessary to extract the file.
If present, the manifest data for the backup is used; if not
present, the path and domain are hashed as in older iOS versions.
}
