% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plist.R
\name{decode_binary_plist}
\alias{decode_binary_plist}
\title{Decode binary plist data}
\usage{
decode_binary_plist(x)
}
\arguments{
\item{x}{A raw vector}
}
\value{
The data stored in the plist. Dictionaries are converted
  to named lists, arrays are converted to lists, and other objects
  are returned as their equivalent atomic types
}
\description{
Decode binary plist data
}
\details{
Some of the databases in the iOS backups contain columns with
binary data that appear to be binary plist objects. This function
can be used to decode the raw data into an proper R object.
}
\examples{
example <- system.file("extdata", "example.bplist", package = "myphone")
bytes <- readBin(example, "raw", n=file.size(example))
decode_binary_plist(bytes)
}
\references{
\url{https://en.wikipedia.org/wiki/Property_list}

\url{http://fileformats.archiveteam.org/wiki/Property_List/Binary}

\url{https://github.com/google/binplist/}
}
