% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backups.R
\name{get_backup}
\alias{get_backup}
\title{Get iOS Backup}
\usage{
get_backup(x = 1)
}
\arguments{
\item{x}{If unspecified, returns the newest back-up. You can also
specify a full path to a backup or a partial name that will be
matched against known backup names}
}
\value{
An object of class \code{ios_backup}
}
\description{
Find iOS backup path
}
