% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backups.R
\name{get_backups}
\alias{get_backups}
\title{Get iOS Backups}
\usage{
get_backups(...)
}
\arguments{
\item{...}{Parameters passed to \code{list_backups}}
}
\value{
A list of all backups in backup folder
}
\description{
Find all iOS backup paths
}
\examples{
\dontrun{
backups <- get_backups()
if (length(backups)) {
  Map(function(x) {x$path}, backups)
}
}
}
