% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sms.R
\name{grep_emoji}
\alias{grep_emoji}
\title{Find emojis in text}
\usage{
grep_emoji(x)
}
\arguments{
\item{x}{A character vector}
}
\value{
A list the same length as \code{x} with a vector
 of emoji-like symbols found in the text
}
\description{
Extract emojis from text
}
