% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backups.R
\name{manifest_contents}
\alias{manifest_contents}
\title{Return manifest contents}
\usage{
manifest_contents(backup, table = "Files", collect = TRUE)
}
\arguments{
\item{backup}{An \code{ios_backup} object (or something that can be passed
to \code{get_backup}).}

\item{table}{Which table should be returned from the manifest database.}

\item{collect}{Should dplyr results be collected before being returned.}
}
\value{
This will return a tibble with the contact data.
  If \code{collect==FALSE}, it will be a lazy tibble.
  The following columns will be included
\itemize{
\item{fileID} A unique ID for each file in the backup
\item{domain} The domain where the file is used
\item{flags} Flags set on the file
}
}
\description{
List all files in a backup from the manifest
}
\details{
A manifest file is a database that tracks all of the files in a backup.
The fileID has traditioanally been a hash of the combination of the
domain and the relative bath but by looking it up in the manifest, we
can be sure to extract the correct file.
}
