% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contacts.R
\name{read_contact_entities}
\alias{read_contact_entities}
\title{Read contacts}
\usage{
read_contact_entities(x, collect = TRUE)
}
\arguments{
\item{x}{An \code{ios_backup} object (or something that can be passed
to \code{get_backup}) or the path to a sqlite AddressBook database.}

\item{collect}{Should dplyr results be collected before being returned.}
}
\value{
This will return a tibble with the contact data.
  If \code{collect==FALSE}, it will be a lazy tibble.
  The following columns will be included
\itemize{
\item{contact_id} A unique ID for each contact
\item{first_name} Contact first name
\item{middle_name} Contact middle name
\item{last_name} Contact last name
\item{organization} Contact organization
\item{department} Contact department
\item{jobtitle} Contact job title
\item{birthday} Contact birthday
\item{creation_date} Date contact created
}
}
\description{
Extract contacts from the Address Book in an iOS Backup.
}
\seealso{
\code{\link{read_contacts_phone_numbers}}, \code{\link{read_contacts_email_addresses}}
}
