% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contacts.R
\name{read_contacts}
\alias{read_contacts}
\title{Read contacts database}
\usage{
read_contacts(x, collect = TRUE)
}
\arguments{
\item{x}{An \code{ios_backup} object (oe somethin that can be passed
to \code{get_backup}) or the path to a sqlite AddressBook database.}

\item{collect}{Should dplyr results be collected before being returned.}
}
\description{
Extract data from the Address Book in an iOS Backup.
}
