% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contacts.R
\name{read_contacts_phone_numbers}
\alias{read_contacts_phone_numbers}
\title{Read contact phone numbers}
\usage{
read_contacts_phone_numbers(x, collect = TRUE)
}
\arguments{
\item{x}{An \code{ios_backup} object (or something that can be passed
to \code{get_backup}) or the path to a sqlite AddressBook database.}

\item{collect}{Should dplyr results be collected before being returned.}
}
\value{
This will return a tibble with the contact data.
  If \code{collect==FALSE}, it will be a lazy tibble.
  The following columns will be included
\itemize{
\item{contact_id} A unique ID for each contact
\item{identifier} A unique number assigned to each address 
   for each contact (0 indicates "primary" address)
\item{phone} Contact phone number
\item{phone_type} Phone number label
}
}
\description{
Extract phone numbers from the Address Book in an iOS Backup.
}
\seealso{
\code{\link{read_contact_entities}}, \code{\link{read_contacts_email_addresses}}
}
