% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sms.R
\name{read_sms_data}
\alias{read_sms_data}
\title{Read SMS Data}
\usage{
read_sms_data(x, collect = TRUE)
}
\arguments{
\item{x}{An \code{ios_backup} object (or something that can be passed
to \code{get_backup}) or the path to a sqlite SMS database.}

\item{collect}{Should dplyr results be collected before being returned.}
}
\value{
A tibble containing SMS data
}
\description{
Return SMS messages stored in backup
}
\examples{
\dontrun{
backup <- get_backup()
read_sms_data(backup)

#Read a copy
orig <- path_sms(backup)
mycopy <- file.path(getwd(), "sms.sqlitedb")
file.copt(orig, mycopy)
read_contacts(mycopy)

#Join with Contacts
read_sms_data(backup) \%>\% left_join(read_contacts(backup))
}
}
