% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sms.R
\name{sms_attachment_path}
\alias{sms_attachment_path}
\title{Find SMS attachment path}
\usage{
sms_attachment_path(backup, attachment_filename)
}
\arguments{
\item{backup}{An \code{ios_backup} object}

\item{attachment_filename}{The attachment filename from the SMS database}
}
\value{
The full path to the attachment file in the backup
}
\description{
Return the file path to an SMS attachment given the relative path
 from the SMS database
}
\examples{
\dontrun{
backup <- get_backup()
sms <- read_sms_data(backup)
sms_attachment_path(backup, na.omit(sms$attachment_filename))
}
}
