% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contacts.R
\name{read_contacts}
\alias{read_contacts}
\title{Read contacts database}
\usage{
read_contacts(x, collect = TRUE)
}
\arguments{
\item{x}{An \code{ios_backup} object (or something that can be passed
to \code{get_backup}) or the path to a sqlite AddressBook database.}

\item{collect}{Should dplyr results be collected before being returned.}
}
\value{
This will return a tibble with the contact data.
  If \code{collect==FALSE}, it will be a lazy tibble.
  The following columns will be included
\itemize{
\item{contact_id} A unique ID for each contact
\item{first} Contact first name
\item{last} Contact last name
\item{emain_phone} Contact e-mail or phone number
}
}
\description{
Extract data from the Address Book in an iOS Backup.
}
\examples{
\dontrun{
backup <- get_backup()
read_contacts(backup)

#Read a copy
orig <- path_contacts(backup)
mycopy <- file.path(getwd(), "contacts.sqlitedb")
file.copt(orig, mycopy)
read_contacts(mycopy)

#Join with SMS
read_sms_data(backup) \%>\% left_join(read_contacts(backup))
}
}
